requirements.txt
streamlit==1.38.0
requests==2.32.3
beautifulsoup4==4.12.3
pymongo==4.8.0
sentence-transformers==3.1.1
python-dotenv==1.0.1
openai==1.51.0
scikit-learn==1.5.1
numpy==2.1.1
